Feature/modify aicore

+++ Image_classfication/s3_access.py
@@ -1,8 +1,10 @@
 import os
 import boto3
+from boto3.s3.transfer import S3Transfer
 from dotenv import load_dotenv
 
 load_dotenv()
+
 
 def get_s3_client():
     credentials = {
@@ -11,6 +13,7 @@ def get_s3_client():
     }
     client = boto3.client("s3", "ap-southeast-1", **credentials)
     return client
+
 
 def get_s3_bucket():
     s3 = boto3.resource(
@@ -22,13 +25,24 @@ def get_s3_bucket():
     return source_bucket
 
 
-def upload_file(bucket, file_path, key_file):
+def upload_file(client, file_path, key_file):
+    transfer = S3Transfer(client)
+    transfer.upload_file(
+        file_path, os.getenv("BUCKET_NAME"), key_file, extra_args={"ACL": "public-read"}
+    )
+    if not file_path.startswith("sample/"):
+        os.remove(file_path)
+
+
+def upload_weights_file(bucket, file_path, key_file):
     bucket.upload_file(file_path, key_file)
-    os.remove(file_path)
+    # os.remove(file_path)
+
 
 def get_public_url(client, key_file):
     file_url = "%s/%s/%s" % (client.meta.endpoint_url, os.getenv("BUCKET_NAME"), key_file)
     return file_url
+
 
 def download_file(bucket, file_name, file_path):
     bucket.download_file(file_name, file_path)
